@@ -38,7 +38,7 @@
     "@popperjs/core": "^2.11.6",
     "@rollup/pluginutils": "^5.0.2",
     "@testing-library/jest-dom": "^5.15.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.4.0",
     "@vant/touch-emulator": "^1.3.2",
     "@vitejs/plugin-react": "^2.0.1",
     "address": "^1.1.2",
@@ -64,7 +64,7 @@
     "ora": "^5.1.0",
     "postcss": "^8.1.7",
     "postcss-load-config": "^3.1.4",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.15.0",
     "release-it": "^15.4.1",
     "rehype-highlight": "^6.0.0",
     "signale": "^1.4.0",
@@ -85,13 +85,13 @@
     "@types/markdown-it": "^12.2.3",
     "@types/node": "^14.11.8",
     "@types/postcss-load-config": "^3.0.1",
-    "@types/react": "^17.0.3",
-    "@types/react-router-dom": "^5.1.7",
+    "@types/react": "^18.2.20",
+    "@types/react-router-dom": "^5.3.3",
     "@types/signale": "^1.4.1",
     "@types/through2": "^2.0.36",
     "@types/vinyl-fs": "^2.4.11",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "release-it": {
     "git": {
