@@ -20,7 +20,7 @@
     "build:site": "rc-cli build-site",
     "release:site": "pnpm build:site && gh-pages -d site",
     "prepack": "pnpm build",
-    "test": "rc-cli test --colors --coverage",
+    "test": "rc-cli test --u --colors --coverage",
     "test:coverage": "open tests/coverage/index.html"
   },
   "author": "rancui",
@@ -62,14 +62,14 @@
   "devDependencies": {
     "@rancui/cli": "workspace:*",
     "@rancui/linter": "workspace:*",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^13.4.0",
     "@types/jest": "^27.0.3",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.2.20",
+    "@types/react-dom": "^18.2.7",
     "@types/react-transition-group": "^4.4.5",
     "gh-pages": "^3.2.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "jest-canvas-mock": "2.5.2"
   },
   "browserslist": [
